+++ test/test/conftest.py
@@ -98,3 +98,15 @@ def iris_targets_df(iris_df: pd.DataFrame, iris_target_name: str) -> pd.DataFram
 @pytest.fixture  # type: ignore
 def iris_targets_binary_df(iris_target_sr: pd.Series) -> pd.DataFrame:
     return OneHotEncoderDF(sparse=False).fit_transform(X=iris_target_sr.to_frame())
+
+
+@pytest.fixture  # type:ignore
+def test_data_categorical() -> pd.DataFrame:
+    return pd.DataFrame(
+        data=[
+            ["yes", "red", "child"],
+            ["yes", "blue", "father"],
+            ["no", "green", "mother"],
+        ],
+        columns=["a", "b", "c"],
+    )